1. use hooks 2. refactor init
