+++ frontend/src/hooks/mod.rs
@@ -1,0 +1,2 @@
+mod initialize_files;
+pub use initialize_files::*;
