+++ frontend/src/lib.rs
@@ -5,6 +5,9 @@ extern crate yew;
 
 use std::env;
 
+pub mod hooks;
+
+// use hooks::*;
 use lazy_static::lazy_static;
 use wasm_bindgen::prelude::*;
 
@@ -17,7 +20,6 @@ mod components;
 mod router;
 mod test;
 mod utils;
-
 
 
 #[wasm_bindgen]
